refactor: 디렉토리구조 변경 및 파일명변경
