+++ src/app/page.js
@@ -3,7 +3,7 @@
 import { HeroSection } from "./_components/sections/heroSection";
 import { ContactSection } from "./_components/sections/contactSection";
 import { ServiceSection } from "./_components/sections/serviceSection";
-import { FAQSection } from "./_features/sections/faqSection";
+import { FAQSection } from "./_components/sections/faqSection";
 
 export default function Home() {
   return (
